# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "version": "1.0.0",
      "description": "API for Social media",
      "title": "Social-Media API"
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      }
    ],
    "components": {
      "securitySchemes": {
        "JWT": {
          "type": "apiKey",
          "in": "header",
          "name": "Authorization",
          "description": "Enter your JWT token here (e.g., 'Bearer your-token')"
        }
      }
    },
    "paths": {
...
      "/api/posts/":{
        "get": {
          "tags": ["Posts"],
          "description": "User will get all the posts posted by him",
          "security": [
            {
              "JWT": []
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "400": {
              "description": "Not Found"
            }
          }
        },
        "post": {
            "tags": ["Posts"],
          "summary": "User can add a new post",
          "description": "Logged-in users can add a new post",
          "requestBody": {
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "caption": {
                      "type": "string",
                      "description": "This is the caption of the post"
                    },
                    "image": {
                      "type": "string",
                      "format": "binary",
                      "description": "This is the image file to upload"
                    }
                  },
                  "required": ["caption", "image"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            },
            "400": {
              "description": "Bad Request"
            }
          }
        }
        
      },
...
   }
}